+++ api/send-email.js
@@ -1,0 +1,3 @@
+import handler from '../functions/send-email.js';
+
+export default handler;